Add wall detection collision for the Player. Duplicate code for wall collision for both the Roamer class and the Player class. This makes me think this needs to be abstracted away, but I am not sure where to? Possibly the rectangle class?
Also found an interesting edge case when the player size was the same as
the tile size of the walls, this lead to not being able to slip through
doorways so my temporaray solution/hack was to make the player and
roamer smaller.

All of this is just more proof of concepts and will probably ultimately
go away, this is more of a learning experience.